Update Dockerfile to work standalone.
In general users will use the separate OpenEats repository which defines the
docker-compose file necessary to run an entire install together. This is fine,
but some advanced users may want to run the containers separately. Without these
changes it isn't clear which ports are exposed and the container does not
actually do anything.